hardware: qcom: sensors: Fix for SHD restart 90 seconds delay
If we use a return statement in the signal handler function, 
the function will return to the point immediately following the 
point where the signal was received. However, the signal itself 
will not be cleared and the signal may be delivered again, causing 
the signal handler to be invoked again.

This can result in an infinite loop of signal handling, which may 
cause the program to crash or behave unexpectedly. In addition, if 
the signal handler modifies any global or static variables, these 
variables may be in an inconsistent state if the signal is delivered 
again.

Therefore, using the exit() function instead of return in a signal 
handler function is recommended to terminate the program immediately 
and avoid any potential issues with signal handling.

Change-Id: I79ce8e51f28a903829890bdd72b0f403da59a230
Signed-off-by: Raghava Chowdam <quic_rchowdam@quicinc.com>