hardware: qcom: sensors: Fix for sensor-hal-daemon crash
In the current implemetation, we're closing the buffer
read files when SHD recieves delete request. This results
in SHD to crash/restart as the data is still being read.

With this change, the buffer read files will be closed
only when all the data is read even when SHD recieves
delete request. This is a graceful handle and SHD will
not crash.

Change-Id: Iebae5715dbcc5e156afecf425f755914c6aad392
Signed-off-by: Raghava Chowdam <quic_rchowdam@quicinc.com>