hardware: qcom: sensors: Fix sensor batching support algorithm issue
1. Resolved the rebatching algorithm issue by accommodating all possible
batch rates for the requested sampling rate. Initially, it ranged from
minBatchCount to maxBatchCount, but now it includes all possible batch
rates for the given requested sampling rate and the physically
configured sensor rate. For example, for 400 & 20, the supported minimum
batch count is 5 for 100Hz and 10 for 200Hz, whereas it was initially 20
for both 100Hz and 200Hz.
2. Implemented logic to support nearby sampling rates for the requested
sampling rate, either from /etc/sensors.conf or the application.
3. Added a blocking call in the sensor test app to prevent high CPU
usage.

Change-Id: Ib618896476ff3ac976f0e499b2c458c76a2fef26