hardware: qcom: sensors: fix the failure message
When systemd shuts down, it sends a SIGTERM signal to all services.
SHD handles this signal with a signal handler, which calls the
destructor to gracefully clean up and free all resources.

However, while exiting the signal handler, SHD uses the received signal
to call the exit function, which results in the shutdown being
considered a failure, even though SHD shuts down gracefully.
Now, instead of using the signal, SHD passes 0 to the exit function,
indicating a successful shutdown.

Change-Id: I9dd5f74e60072c001d15cd36b4fe5d57ceaa536d
Signed-off-by: Raghava Chowdam <quic_rchowdam@quicinc.com>