hardware: qcom: sensors: fix for shd crash during bootup
Initialize the IPC and Qsocket interfaces to nullptr to
ensure that, during the conditional check in the destructor,
these pointers are not assigned garbage values. This will also
allow the listener function in the socket library to execute
correctly and avoid SHD crash incase of opening the HAL library
is failed.

Change-Id: I3efb83cd13e423ff2ebba47f8f2a696d6fed5437
Signed-off-by: Raghava Chowdam <quic_rchowdam@quicinc.com>