hardware: qcom: sensors: fix the failure message
When systemd shuts down, it sends a SIGTERM signal to all services.
SHD handles this signal with a signal handler, which calls the
destructor to gracefully clean up and free all resources.

However, while exiting the signal handler, SHD uses the received signal
to call the exit function, which results in the shutdown being
considered a failure, even though SHD shuts down gracefully.
Now, instead of using the signal, SHD passes 0 to the exit function,
indicating a successful shutdown.

Change-Id: I476dcdec8e023f9091de3ca0e6bbfdc9f265e5b5
Signed-off-by: Puneet Yatnal <quic_puneety@quicinc.com>