sensors: ASM hal config thread to ignore config file
ASM330 hal_configuration_thread configures the HAL and
writes it to the file. If the file is not present the hal
config is also skipped. This change ensures hal config is updated
even if hal config file is not present. If the file is present,
it write the config to the file additionally.

Change-Id: I00bac2f14276842fc3bacc3c10dd3e87e72e9239
(cherry picked from commit d85b4eea6705c2a149aabb93613aec63c41e58a5)