hardware: qcom: sensors: Disable wakeup on client exit
Disable wakeup on motion when the client exits
gracefully. If a client that has enabled wom exits, and
there is no other client with wom enabled, we should
disable the wom. No doing so, SHD assumes wom is enabled
and doesn't allow disabling / reenabling when another
client tires to do so.

Change-Id: I00dc0a585f94d7c1348c5287270b2cdbe3b8768d
CRs-Fixed: 4307438